Replace 2025-04-23 measurements with 2025-04-09 ones
The 2025-04-23 measurements were carried away using calibration
with the calibxyzkins module. This means they cannot be used to
analyze the error of the uncalibrated gantry robot. Moreover
the results are incorrect because we made a mistake and used
incorrect calibration parameters.

The 2025-04-09 measurements were carried away without calibration,
thus this is the baseline dataset for error analysis and evaluation
of the calibration effectiveness.